Re-routed shelters page to show all shelters, shelter card to open shelter page
