+++ client/src/App.js
@@ -13,10 +13,11 @@ import { DialogsProvider } from "./utils/contexts/DialogsContext";
 import DialogsContainer from "./components/AccountDialogs";
 
 import Home from "./pages/Home";
-import Shelters from "./pages/Shelters";
+import Shelter from "./pages/Shelter";
 import Header from "./components/Header";
 import Nav from "./components/Nav";
 import Success from "./pages/Success";
+import Shelters from "./pages/Shelters";
 
 // Create an HTTP link to connect to the GraphQL server -> The link is configured to send requests to the "/graphql" endpoint.
 // The "uri" option specifies the URL to which the requests will be sent.
@@ -86,6 +87,7 @@ function App() {
           <Routes>
             <Route path="/" element={<Home />} />
             <Route path="/Shelters" element={<Shelters />} />
+            <Route path="/Shelters/:Id" element={<Shelter />} />
             <Route
               path="*"
               element={<h1 className="display-2">Wrong page!</h1>}
