+++ client/src/pages/Home.js
@@ -1,12 +1,10 @@
 import React from "react";
 import Hero from "../components/Hero";
-import AllShelters from "./AllShelters";
 
 const Home = () => {
   return (
     <div className="Home">
       <Hero />
-      <AllShelters />
     </div>
   );
 };
